(#1409) Improve comments and minor structure cleanup
