+++ external_scripts/NRT/Saildrone_conversion/saildrone_main.py
@@ -1,10 +1,10 @@
 ###############################################################################
-### SAILDRONE DATA RETRIEVAL AND CONVERSION		                            ###
+### THE MAIN SAILDRONE SCRIPT                                               ###
 ###############################################################################
 
 ### Description
-# Script uses Saildrone API to download saildrone ocean and biogeo data,
-# converts it from json to csv format, and shares it with QuinCe.
+# Script uses Saildrone API to download saildrone data, converts it from json
+# to csv format, and shares it with QuinCe.
 
 
 ###----------------------------------------------------------------------------
@@ -15,9 +15,7 @@ import os
 import saildrone_module as saildrone
 from datetime import datetime
 import shutil
-import urllib
 import json
-import ast
 import pandas as pd
 
 
@@ -36,7 +34,6 @@ if not os.path.isdir('./data_files'):
 data_dir = os.path.join(script_dir,'data_files')
 
 # Create new archive directry with current timestamp as name
-# !!! Can remove this when everything works
 now = datetime.now()
 dt_string = now.strftime("%Y%m%dT%H%M%S")
 archive_path = os.path.join(data_dir, str(dt_string))
@@ -44,7 +41,7 @@ os.mkdir(archive_path)
 
 
 ###----------------------------------------------------------------------------
-### Extract information from the config and stored info files
+### Extract information from the config and stored_info files
 ###----------------------------------------------------------------------------
 
 try:
@@ -78,25 +75,22 @@ except FileNotFoundError:
 token = saildrone.auth()
 access_list = saildrone.get_available(token)
 
-# Check if the access list has changed since the previous run (and previous
-# access list was not empty): send message and replace the prev_access_list
+# If the access list has changed since the previous run (and previous
+# access list was not empty): print message and replace the prev_access_list
 # with the new access_list.
 if prev_access_list != access_list and bool(prev_access_list):
 	#!!! Send message to slack. Temp soluion:
-	print("Access_list has changed")
+	print("Access list has changed")
 	stored_info['prev_access_list'] = access_list
 
-# If the access list contains new drones which are not on ignore list, add
+# If the access list contains new drones which are not on the ignore list, add
 # them to the next_request dictionary:
 for dictionary in access_list:
 	drone = str(dictionary['drone_id'])
 	if drone not in drones_ignored and drone not in next_request.keys():
 			next_request[drone] = dictionary['start_date']
 
-# Function 'check_next_request' will return a list of next requests where
-# drone items are removed if they are on the ignore list, OR if any of the
-# following is not available: the drone itself, any of the dataset typs, or the
-# start date.
+# Function 'check_next_request' will return what we need to request.
 next_request_checked = saildrone.check_next_request(
 	next_request, access_list, datasets, drones_ignored)
 
@@ -111,8 +105,8 @@ end = now.strftime("%Y-%m-%dT%H:%M:%S") + ".000Z"
 # Create connection to the Quince FTP
 ftpconn = saildrone.connect_ftp(FTP)
 
-# Loop that downloads, converts, merges, writes datafiles and send it to the
-# QuinCe ftp site. Keep track on next start requests in next_request_updated.
+# Loop that downloads, converts, merges, and sends data files it to the QuinCe
+# FTP. Keep track on what to request next time in the next_request_updated.
 next_request_updated = dict(next_request_checked)
 for drone_id, start in next_request_checked.items():
 
@@ -123,8 +117,8 @@ for drone_id, start in next_request_checked.items():
 			data_dir, drone_id, dataset, start, end, token)
 		json_paths.append(json_path)
 
-	# Convert each json to csv. Move the json file to the archive folder.
-	# Store the new csv paths.
+	# Convert each json to csv. Move the json file to the archive folder. Store
+	# the paths of the csv files.
 	csv_paths = []
 	for path in json_paths :
 		csv_path = saildrone.convert_to_csv(path)
@@ -134,25 +128,27 @@ for drone_id, start in next_request_checked.items():
 	# Create merged dataframe
 	merged_df = saildrone.merge_datasets(csv_paths)
 
-	# Add missing columns to ensure consistent data format:
+	# Move the individual csv files to archive
+	for path in csv_paths:
+		shutil.move(path, os.path.join(archive_path,
+		os.path.basename(path)))
+
+	# Add missing columns (with empty data), and sort by the defined column
+	# order to ensure consistent data format
 	for param in col_order:
 		if param not in merged_df.columns:
 			merged_df[param] = None
-
 	# !!! Check: If there are new headers in the merged_df.
 	# Notify slack and stop script
-
-	# Sort by the defined column order from the config file
 	merged_sorted_df = merged_df[col_order]
 
 	# Get the last record we downloaded from the biogeo dataset. This will
 	# change once the different SailDrone datasets are no longer merged
-	# together. This will be used as the starting point for the next
-	# request.
-	col_index = merged_sorted_df.columns.get_loc('time_interval_biogeFile')
-	last_record_date = merged_sorted_df.tail(1).iloc[0,col_index]
+	# together. This will be used as the starting point for the next request.
+	time_index = merged_sorted_df.columns.get_loc('time_interval_biogeFile')
+	last_record_date = merged_sorted_df.tail(1).iloc[0,time_index]
 
-	# Export the merged data to a csv file
+	# Store the merged data as a csv file in the archive folder
 	merged_file_name = (str(drone_id) + '_'
 		+ start[0:4] + start[5:7] + start[8:10] + 'T' + start[11:13]
 		+ start[14:16] + start[17:19] + "-"
@@ -161,23 +157,24 @@ for drone_id, start in next_request_checked.items():
 	merged_csv = merged_sorted_df.to_csv(merged_path,
 		index=None, header=True, sep=',')
 
-	# Move the individual csv files to archive
-	for path in csv_paths:
-		shutil.move(path, os.path.join(archive_path,
-		os.path.basename(path)))
-
-	# Export the merged dataset to the Quince FTP site
-	upload_result = saildrone.upload_file(ftpconn=ftpconn,
-		ftp_config=FTP, instrument_id=1000, filename=merged_file_name,
-		contents=merged_csv)
+	# Open the merged csv file in byte format and send to the Quince FTP
+	with open(merged_path, 'rb') as file:
+		byte = file.read()
+		upload_result = saildrone.upload_file(ftpconn=ftpconn,
+			ftp_config=FTP, instrument_id=1000, filename=merged_file_name,
+			contents=byte)
 
 	#  Set new start date for the next_request:
 	next_request_updated[drone_id] = (last_record_date
 		+ pd.Timedelta("1 minute")).strftime("%Y-%m-%dT%H:%M:%S.000Z")
 
 
+###----------------------------------------------------------------------------
+### Prepare for next download request
+###----------------------------------------------------------------------------
+
 # Update stored_info file
-stored_info['next_request'] = next_request_updated
-with open('./stored_info.json', 'w') as file:
-	json.dump(stored_info, file,
-		sort_keys=True, indent=4, separators=(',',': '))
+#stored_info['next_request'] = next_request_updated
+#with open('./stored_info.json', 'w') as file:
+#	json.dump(stored_info, file,
+#		sort_keys=True, indent=4, separators=(',',': '))
