+++ external_scripts/NRT/Saildrone_conversion/saildrone_module/__init__.py
@@ -1,6 +1,7 @@
 from .convert import convert_to_csv
 from .merge import merge_datasets
-from .api import to_dict, write_json, auth, get_available, check_next_request
+from .api import to_dict, write_json, auth, get_available
+from .check_request import check_next_request
 from .nrtftp import connect_ftp, upload_file
 
 __all__ = ['convert_to_csv', 'merge_datasets', 'to_dict', 'write_json', 'auth',
