+++ external_scripts/NRT/Saildrone_conversion/saildrone_module/convert.py
@@ -12,14 +12,15 @@ import json
 def convert_to_csv(file_path):
 
 	# Read the json file and create the 'data_list' variable. This variable
-	# contains a list of dictionaries. Each dictionary contains several
-	# parameter-value pairs, always including one timestamp parameter-value.
+	# contains the actual data stored as a list of dictionaries. Each
+	# dictionary contains data from several parameters from one timestamp.
 	with open(file_path) as json_file:
 		json_loaded = json.load(json_file)
 	data_list = json_loaded['data']
 
-	# The parameters can only contain one value per timestamp/dictionary.
-	# Must therefore remove parameters containing a list of values.
+	# Sometimes parameters are lists of datapoints, but we can only allow one
+	# data point per timestamp (per dictionary). Must therefore remove
+	# parameters containing a list of values.
 	data_list_edited = []
 	for dictionary in data_list:
 
@@ -33,16 +34,17 @@ def convert_to_csv(file_path):
 		if len(dict_to_keep) > 1:
 			data_list_edited.append(dict_to_keep)
 
-	# Find distinct parameters.
+	# Find distinct parameters. These are then used as headers in the csv file.
 	distinct_parameters = []
 	for dictionary in data_list_edited:
 		for parameter in dictionary.keys():
 			if parameter not in distinct_parameters:
 				distinct_parameters.append(parameter)
 
-	# Create output file name and write to the file
+	# Create output csv file name and write to the file
 	csv_file_path = file_path.replace('.json','') + ".csv"
 	with open(csv_file_path, 'w', newline = '') as output_csv:
+		# The csv.DictWriter function converts dictionaries to csv format.
 		output_writer = csv.DictWriter(
 			output_csv, fieldnames = distinct_parameters)
 		output_writer.writeheader()
